backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.1
boto3==1.29.3
botocore==1.32.7
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.4.2
numpy==1.26.2
packaging==24.2
pandas==2.1.3
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.1
pydantic_core==2.14.3
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.2
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.7
uvicorn==0.24.0
